+++ pyraxshell/plugins/plugin_auth.py
@@ -17,11 +17,11 @@
 
 import cmd
 import logging
-from utility import kvstring_to_dict
+import traceback
+
+from globals import ERROR, INFO
 from plugins.libauth import LibAuth
 from plugins.plugin import Plugin
-from globals import INFO, ERROR
-import traceback
 
 name = 'auth'
 
@@ -110,46 +110,34 @@ class Cmd_auth(Plugin, cmd.Cmd):
         
         Parameters:
         
-        identity_type = 'rackspace'    (default)
-        username
         apikey
-        region                         (default: pyrax.default_region)
+        username
+        identity_type    (default: rackspace)
+        region           (default: pyrax.default_region)
         '''
-        logging.debug("line: %s" % line)
-        d_kv = kvstring_to_dict(line)
-        logging.debug("kvs: %s" % d_kv)
-        # default values
-        _identity_type = 'rackspace'
-        _username = None
-        _apikey = None
-        _region = self.libplugin.default_region()
-        # parsing parameters
-        if 'identity_type' in d_kv.keys():
-            _identity_type = d_kv['identity_type']
-        else:
-            logging.debug("identity_type: %s (default)" % _identity_type)
-        if 'username' in d_kv.keys():
-            _username = d_kv['username']
-        else:
-            cmd_out = "missing username"
-            self.r(1, cmd_out, ERROR)
+        # check and set defaults
+        retcode, retmsg = self.kvargcheck(
+            {'name':'apikey', 'required':True},
+            {'name':'username', 'required':True},
+            {'name':'identity_type', 'default':'rackspace'},
+            {'name':'region', 'default':self.libplugin.default_region()}
+        )
+        if not retcode:             # something bad happened
+            self.r(1, retmsg, ERROR)
             return False
-        if 'apikey' in d_kv.keys():
-            _apikey = d_kv['apikey']
-        else:
-            cmd_out = "missing apikey"
-            self.r(1, cmd_out, ERROR)
-            return False
-        if 'region' in d_kv.keys():
-            _region = d_kv['region']
+        self.r(0, retmsg, INFO)     # everything's ok
+        
         try:
-            self.libplugin.authenticate_login(identity_type = _identity_type,
-                                                  username = _username,
-                                                  apikey = _apikey,
-                                                  region = _region)
+            self.libplugin.authenticate_login(
+                apikey = self.kvarg['apikey'],
+                username = self.kvarg['username'],
+                identity_type = self.kvarg['identity_type'],
+                region = self.kvarg['region'],
+            )
             cmd_out  = ('login - indentity_type:%s, username=%s, apikey=%s, '
-                        'region=%s' %
-                        (_identity_type, _username, _apikey, _region))
+                        'region=%s' % 
+                        (self.kvarg['identity_type'], self.kvarg['username'],
+                         self.kvarg['apikey'], self.kvarg['region']))
             self.r(0, cmd_out, INFO)
         except:
             tb = traceback.format_exc()
@@ -186,52 +174,30 @@ class Cmd_auth(Plugin, cmd.Cmd):
         
         Parameters:
         
-        identity_type
-        region
+        identity_type    (default: rackspace)
+        region           (default: LON)
         tenantId
         token
         '''
-        logging.debug("line: %s" % line)
-        d_kv = kvstring_to_dict(line)
-        logging.debug("kvs: %s" % d_kv)
-        # default values
-        _identity_type = None
-        _region = None
-        _tenantId = None
-        _token = None
-        if 'identity_type' in d_kv.keys():
-            _identity_type = d_kv['identity_type']
-        else:
-            cmd_out = "missing identity_type"
-            self.r(1, cmd_out, ERROR)
+        retcode, retmsg = self.kvargcheck(
+              {'name':'identity_type', 'default':'rackspace'},
+              {'name':'region', 'default':'LON'},
+              {'name':'tenantId', 'required':True},
+              {'name':'token', 'required':True}
+        )
+        if not retcode:
+            self.r(1, retmsg, ERROR)
             return False
-        if 'region' in d_kv.keys():
-            _region = d_kv['region']
-        else:
-            cmd_out = "missing region"
-            self.r(1, cmd_out, ERROR)
-            return False
-        # parsing parameters
-        if 'tenantId' in d_kv.keys():
-            _tenantId = d_kv['tenantId']
-        else:
-            cmd_out = "missing tenantId"
-            self.r(1, cmd_out, ERROR)
-            return False
-        if 'token' in d_kv.keys():
-            _token = d_kv['token']
-        else:
-            cmd_out = "missing token"
-            self.r(1, cmd_out, ERROR)
-            return False
-        #
-        print "-" * 10
-        print _tenantId, _token
+        self.r(0, retmsg, INFO)
         try:
-            self.libplugin.authenticate_token(token=_token, tenantId=_tenantId,
-                                              region=_region,
-                                              identity_type=_identity_type)
-            cmd_out = 'login - tenantId=%s, token=%s' % (_tenantId, _token)
+            self.libplugin.authenticate_token(
+                identity_type=self.kvarg['identity_type'],
+                region=self.kvarg['region'],
+                tenantId=self.kvarg['tenantId'],
+                token=self.kvarg['token'],
+            )
+            cmd_out = ('login - tenantId=%s, token=%s' %
+                       (self.kvarg['tenantId'], self.kvarg['token']))
             self.r(0, cmd_out, INFO)
         except:
             tb = traceback.format_exc()